Split test files
My test cases are long and scrolling is annoying
