+++ tox.ini
@@ -22,7 +22,7 @@ exclude=
     venv,.venv,
     .tox
 per-file-ignores =
-    tests/test_reflow.py:E501
+    tests/test_*.py:E501
 
 [pytest]
 minversion = 6.0
